Allow classifications of existing polygons.

+++ imports/ie/ui/components/widgets/MapboxGLDraw.jsx
@@ -22,7 +22,8 @@ export default class MapboxGLDraw extends Component {
 				this.state.draw.changeMode(MAPBOX_MODES.STATIC);
 				break;
 			case DRAWING_STATES.SWITCHING:
-				this.state.draw.deleteAll();
+				if(this.state.draw)
+					this.state.draw.deleteAll();
 				break;
 		}
 	}
@@ -78,7 +79,8 @@ export default class MapboxGLDraw extends Component {
 			switch (e.mode) {
 				case MAPBOX_MODES.SIMPLE_SELECT:
 					thiz.setFeatureAsPolygon(feature);
-					thiz.state.draw.deleteAll();
+					if(thiz.state.draw)
+						thiz.state.draw.deleteAll();
 					break;
 			}
 		});
